Rationalise locations
