+++ kymata/plot/__init__.py
@@ -1,4 +1,4 @@
-from .plot import expression_plot
+from .expression import expression_plot
 
 __all__ = [
     "expression_plot",
